+++ wdl/model/draft2/src/main/scala/wdl/draft2/model/WdlRuntimeAttributes.scala
@@ -12,8 +12,11 @@ case class WdlRuntimeAttributes(attrs: Map[String, WdlExpression]) {
     // In future WDL versions, `container` has superceded `docker` as the runtime attribute for specifying a
     // container image. For pre-1.1 WDLs, we need to remove `container` if it exists so that it doesn't interfere
     // with `docker`.
+    // Similar deal with `gpu` which was introduced in WDL 1.1. This boolean attr controls a GPU requirement check.
     RuntimeAttributes(
-      attrs.filterNot(m => m._1 == RuntimeAttributesKeys.ContainerKey).safeMapValues(WdlWomExpression(_, task))
+      attrs
+        .filterNot(m => List(RuntimeAttributesKeys.ContainerKey, RuntimeAttributesKeys.GpuRequiredKey).contains(m._1))
+        .safeMapValues(WdlWomExpression(_, task))
     )
 }